+++ automation/templates/__init__.py
@@ -1,7 +1,5 @@
 from .yaml_spec import YamlSpec
 from .powers import Powers
-from .markdown import Markdown
-from .csv import Csv
-from .dot import Dot
+from .bestiary import Bestiary
 
-__all__ = ["Markdown", "Csv", "Dot", "YamlSpec", "Powers"]
+__all__ = ["YamlSpec", "Powers", "Bestiary"]
