+++ docs/src/1_Mechanics/05_Vulnerabilities.md
@@ -1,4 +1,4 @@
-<!-- DEVELOPERS: Please edit corresponding yml -->
+<!-- DEVELOPERS: Please edit corresponding yaml -->
 
 
 # Combat
